requirements.txt
requests
pycryptodome
PyQt5
pandas
matplotlib